Derek Atkins' rpc backend patch.
This patch adds _most_ of the RPC generated files into the repository
because it appears many recent versions of rpcgen for Linux and NetBSD
do not support the -M (Multi-Thread Safe) output of code.  Bug reports
have been submitted, but I'd like to support RPC in the next two years
before this is fixed.

So, this adds the files gncRpc.h, gncRpc_clnt.c, and gncRpc_svc.c.
The file gncRpc_xdr.c is still generated on the fly.  This
unfortunately means that any changes to the .x files mean you must
regenerate at least the .h file and possibly the other .c files, too.

This patch also modifies the .cvsignore to allow those three files
into the repository.


git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@3823 57a11ea4-9604-0410-9ed3-97b8803252fd