[r17646] Fix Tax-related inconsistency in UI
Changes the 'Sensitive' Property to 'No' from 'Yes' in 'account.glade'
for the 'tax_related_button' widget so that the setting is displayed but
cannot be changed in the account edit dialog. Sets the Tooltip
property to "Use Edit->Tax Options to set the tax-related flag and assign
a tax code to this account." in 'account.glade'.

Patch by "J. Alex Aycinena" <alex.aycinena@gmail.com>

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/branches/2.2@17648 57a11ea4-9604-0410-9ed3-97b8803252fd