[r17680] Bug #339433: FIx price quote dates off by one day
This bug is related to all the rest of the date-time issues, but has an
additional twist because of finance-quote.

Somewhere between the fq-helper and the price db, gnucash rolls the quote
through timespecCanonicalDayTime(), wherein it first runs localtime_r(), to
produce, presumably, "2008-11-03 19:00:00" for my Eastern US location. Having
now screwed up the date, gnucash runs gnc_tm_set_day_middle() on it to enter
the price in the pricedb as of noon on the wrong date.

Patch by David Reiser.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/branches/2.2@17736 57a11ea4-9604-0410-9ed3-97b8803252fd