[20647] Bug #650163: Fix Technicolor style that outputs wrong tags <btd> instead of <b><td>
Patch by Bert:

The technicolor stylesheet outputs </btd> tags where what they actually meant
to do was </b></td>. This may render report outputs wrongly in different
browsers and prevents HTML validation.

The proposed patch removes the bold tags as well (for example for the total
rows) since the font and its weight are set in the options (and thus reflected
in the CSS style). There is no need for the report to set the bold face with
the <b> tag.

In addition, the attached patch also closed the <center> tag at the beginning
of the body (also preventing validation). Note that that this tag is deprecated
for newer version of html (but will pass HTML 4.01 Transitional).

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/branches/2.4@20650 57a11ea4-9604-0410-9ed3-97b8803252fd