[r21570]Fix clean targets and DIST_SOURCES
Several subdirectories have a subdirectory "gnucash", which clean tried to remove by including in the CLEANFILES variable in Makefile.am. This doesn't work, because directories can't be removed with rm, they require rm -r. Add a clean-local: target which calls rm -r on the gnucash subdirectory.

Some Makefile.ams used an AC_SUBSTed variable to define an optionally-compiled subdirectory, then included the variable in SUBDIRS and the actual directory name in DIST_SUBDIRS. This breaks the distclean and maintainer-clean targets when the optional directory is built, because it is then processed twice, and there is no Makefile the second time so make halts. Replaced the AC_SUBSTed variables with ones local to the Makefile.am, switched with an AM_CONDITIONAL. That way automake can figure out the dist_subdirs on its own.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/branches/2.4@21576 57a11ea4-9604-0410-9ed3-97b8803252fd