[r21727] Fix "imported from both (sw_engine) and (gnucash engine)" errors from Guile
The GncNumericErrorCode enums are re-defined in
src/engine/gnc-numeric.scm -- the SWIG way of representing them produces
type errors in the report code. This may well be a problem with the
calling code, but it's easier at this point to just not import the
symbols into sw_engine.

gnc_commodity_table_get_quotable_commodities is explicitly wrapped in
app-utils.i, so we need to not have the default wrapper in sw_engine.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/branches/2.4@21728 57a11ea4-9604-0410-9ed3-97b8803252fd