[r21647]Remove the obsolescent src/doc/design directory from recursive build.
Prevents make distcheck from building the ps and pdf versions of the
texinfo, which imposed a dependency on TeX.

Removing the directory from the dist requires removing the Makefile from
configure, which completely disables building, so it remains. To build
the info file, cd src/doc/design and make info.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/branches/2.4@21729 57a11ea4-9604-0410-9ed3-97b8803252fd