[22067] Remove potentially abandoned file (that has a GPLv2-only license, too).
The last contributors here, Josh Sled and Chris Lyttle, have put their
code under "GPLv2 and GPLv3" license, see r22063, but I'm too lazy to
look up the copyright holders on the file as a whole so I'd rather remove it.

If someone wants to add a similar packaging file again, feel free to do
so, but please don't use GPLv2-only code to do so.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/branches/2.4@22069 57a11ea4-9604-0410-9ed3-97b8803252fd