Further register rewrite work.
This patch is an updated patch which I have been fighting with the sort
model on. It allows you to add, delete and modify basic transactions
and splits with the toolbar buttons. Jump and move to blank are
also working.

Visual indication of read only and future transactions are
also implemented.

As before, do not open same account in old and new register.

Tab key works manually but needs more work, cell editable value does not
seem to be reliable which was the way I was going to auto step over
uneditable cells.

There is another problem if you enter the debit/credit value before the
account, you will get an error in the trace file but entries
are recorded, not sure of answer.

New transactions are not added in date order at the moment, this was
going to be done by the sort model or the existing method of reload
register via the commented out watchers in ledger-display.

All the numeric functions are the existing ones from the rewrite and as
such I have not proved they are correct but a basic transaction seems to
work.

Author: Robert Fewell <14ubobit@gmail.com>

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@22448 57a11ea4-9604-0410-9ed3-97b8803252fd