Minor changes to indicate svn is not the only/primary vcs used.
The code actually has provisions for svn, svk, git and bzr. Only svn and
git are actively used by the core developers though.

This involves changing
- filenames that contained svn while meant for multiple vcs's
- variable and macro names that are relevant for multiple vcs's
- the various locations where the vcs revision info is displayed
  to the user, to mention the actual vcs used instead of standard "svn"

While at the last item, I also changed the text printed when gnucash is
called with the --version parameter, to reuse the translatable string
that is also used for the about dialog. This saves us one string to
translate.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@22475 57a11ea4-9604-0410-9ed3-97b8803252fd