[r22190]Since 2.32, GCache is deprecated and should be replaced by GHashTable.
In the new string cache (split out from qofutil.c/.h), the GHashTable has the
string as key and a small heap-allocated guint refcount as data.  The value is
allocated rather than being used by value because there is no way to just
modify the data field of an existing key.  The key would need to be changed at
the same time, which would free the old key value and invalidate all gchar
pointers of objects which have refs to the key.
 (Unit tests removed, they don't work in the 2.4 branch because the testing infrastructure isn't in place.)

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/branches/2.4@22537 57a11ea4-9604-0410-9ed3-97b8803252fd